+++ test/lollipop/components/rules_test.clj
@@ -1,7 +1,7 @@
-(ns proxy-protect.components.rules-test
+(ns lollipop.components.rules-test
   (:require [clojure.test :refer :all]
             [clojure.tools.logging.test :refer [with-log]]
-            [proxy-protect.components.rules :as rules]))
+            [lollipop.components.rules :as rules]))
 
 (deftest http-method-allowed?-test
   (is (true? (rules/http-method-allowed? ["any"] :post)))
@@ -11,12 +11,12 @@
 
 (deftest valid-rule?-test
   (with-log
-    (is (false? (rules/valid-rule? {:type "absolute"})))
-    (is (false? (rules/valid-rule? {:type "absolute"
+    (is (false? (rules/valid-rule? {:type "exact"})))
+    (is (false? (rules/valid-rule? {:type "exact"
                                     :allowedMethods ["get"]
                                     :source "some-source"
                                     :destination "http://some-destination/endpoint"})))
-    (is (true? (rules/valid-rule? {:type "absolute"
+    (is (true? (rules/valid-rule? {:type "exact"
                                    :allowedMethods ["get"]
                                    :source "/some-source"
                                    :destination "http://some-destination/endpoint"})))))